refactor contact Service into PingStats on Endpoint
