+++ storagenode/contact/endpoint.go
@@ -5,6 +5,7 @@ package contact
 
 import (
 	"context"
+	"sync"
 	"time"
 
 	"go.uber.org/zap"
@@ -14,19 +15,27 @@ import (
 
 	"storj.io/storj/pkg/identity"
 	"storj.io/storj/pkg/pb"
+	"storj.io/storj/pkg/storj"
 )
 
 // Endpoint implements the contact service Endpoints
 type Endpoint struct {
-	log     *zap.Logger
-	service *Service
+	log       *zap.Logger
+	PingStats *PingStats
+}
+
+// PingStats contains information regarding who and when the node was last pinged
+type PingStats struct {
+	mu               sync.Mutex
+	lastPinged       time.Time
+	whoPingedNodeID  storj.NodeID
+	whoPingedAddress string
 }
 
 // NewEndpoint returns a new contact service endpoint
-func NewEndpoint(log *zap.Logger, service *Service) *Endpoint {
+func NewEndpoint(log *zap.Logger) *Endpoint {
 	return &Endpoint{
-		log:     log,
-		service: service,
+		log: log,
 	}
 }
 
@@ -42,6 +51,22 @@ func (endpoint *Endpoint) Ping(ctx context.Context, req *pb.ContactPingRequest) 
 		return nil, status.Error(codes.Unauthenticated, err.Error())
 	}
 	endpoint.log.Debug("pinged", zap.Stringer("by", peerID.ID), zap.Stringer("srcAddr", p.Addr))
-	endpoint.service.wasPinged(time.Now(), peerID.ID, p.Addr.String())
+	endpoint.PingStats.WasPinged(time.Now(), peerID.ID, p.Addr.String())
 	return &pb.ContactPingResponse{}, nil
+}
+
+// WhenLastPinged returns last time someone pinged this node.
+func (stats *PingStats) WhenLastPinged() (when time.Time, who storj.NodeID, addr string) {
+	stats.mu.Lock()
+	defer stats.mu.Unlock()
+	return stats.lastPinged, stats.whoPingedNodeID, stats.whoPingedAddress
+}
+
+// WasPinged notifies the service it has been remotely pinged.
+func (stats *PingStats) WasPinged(when time.Time, srcNodeID storj.NodeID, srcAddress string) {
+	stats.mu.Lock()
+	defer stats.mu.Unlock()
+	stats.lastPinged = when
+	stats.whoPingedNodeID = srcNodeID
+	stats.whoPingedAddress = srcAddress
 }
